lookup/requirements.txt
face_recognition==1.3.0
faiss-gpu==1.7.2
numpy==1.26.4
pillow==10.2.0
flask==3.0.2
retina-face==0.0.17
opencv-python-headless==4.9.0.80
tensorflow==2.15.0
torch==2.2.1
selenium==4.18.1
requests==2.31.0
urllib3==2.2.1
werkzeug==3.0.1
